# Conflicts:
#   web/vue.config.js

+++ dss-data-governance/dss-data-asset-server/src/main/java/com/webank/wedatasphere/dss/data/asset/dao/MetaInfoMapper.java
@@ -15,7 +15,7 @@ import java.util.List;
 public interface MetaInfoMapper {
     Long getTableStorage() throws SQLException;
     List<HiveStorageInfo> getTop10Table() throws SQLException;
-    int getTableInfo(@Param("dbName") String dbName,@Param("tableName") String tableName,@Param("isPartTable") Boolean isPartTable) throws  SQLException;
+    long getTableInfo(@Param("dbName") String dbName,@Param("tableName") String tableName,@Param("isPartTable") Boolean isPartTable) throws  SQLException;
 
     List<HivePartInfo> getPartInfo(@Param("dbName") String dbName, @Param("tableName") String tableName) throws SQLException;
